🔨 refactor(app): move report post to admin feature
